server/opinion/api-1/requirements.txt
PyJWT
azure-functions
azure-data-tables
requests